requirements.txt
Flask
pdfplumber
pymupdf
pandas
python-dateutil
pillow
pytesseract
openpyxl
gunicorn
